move Base method tests to test_base.py
