+++ tests/test_downloader/test_executors/test_athena.py
@@ -69,108 +69,6 @@ class AthenaExecutorTestCase(TestCase):
             ])
 
     #
-    # COMPILE_TO_QUERY
-    #
-    def test_compile_to_query__columns_select(self):
-
-        a = ef.account()
-        d = DownloadRequest.objects.create(
-            created_by=a,
-            spec={
-                'columns': ['product', 'available'],
-                'filters': [],
-            },
-            catalogue_item=self.ci)
-
-        assert normalize_query(self.executor.compile_to_query(d)) == (
-            'SELECT product, available FROM lakey.shopping')
-
-    def test_compile_to_query__numerical_filters(self):
-
-        a = ef.account()
-        d = DownloadRequest.objects.create(
-            created_by=a,
-            spec={
-                'columns': ['product', 'quantity'],
-                'filters': [
-                    {
-                        'name': 'price',
-                        'operator': '>=',
-                        'value': 78.1,
-                    },
-                ],
-            },
-            catalogue_item=self.ci)
-
-        assert self.executor.compile_to_query(d) == (
-            'SELECT product, quantity FROM lakey.shopping '
-            'WHERE price >= 78.1')
-
-    def test_compile_to_query__categorical_filters(self):
-
-        a = ef.account()
-        d = DownloadRequest.objects.create(
-            created_by=a,
-            spec={
-                'columns': ['product', 'quantity'],
-                'filters': [
-                    {
-                        'name': 'product',
-                        'operator': '=',
-                        'value': 'jeans',
-                    },
-                ],
-            },
-            catalogue_item=self.ci)
-
-        assert self.executor.compile_to_query(d) == (
-            'SELECT product, quantity FROM lakey.shopping '
-            'WHERE product = \'jeans\'')
-
-    def test_compile_to_query__many_filters(self):
-
-        a = ef.account()
-        d = DownloadRequest.objects.create(
-            created_by=a,
-            spec={
-                'columns': ['product', 'available'],
-                'filters': [
-                    {
-                        'name': 'price',
-                        'operator': '>=',
-                        'value': 289.2,
-                    },
-                    {
-                        'name': 'product',
-                        'operator': '=',
-                        'value': 'jeans',
-                    },
-                ],
-            },
-            catalogue_item=self.ci)
-
-        assert self.executor.compile_to_query(d) == (
-            'SELECT product, available FROM lakey.shopping '
-            'WHERE price >= 289.2 AND product = \'jeans\'')
-
-    def test_compile_to_query__radomization_filters(self):
-
-        a = ef.account()
-        d = DownloadRequest.objects.create(
-            created_by=a,
-            spec={
-                'columns': ['product', 'available'],
-                'filters': [],
-                'randomize_ratio': 0.9,
-            },
-            catalogue_item=self.ci)
-
-        assert normalize_query(self.executor.compile_to_query(d)) == (
-            'SELECT q.* FROM ('
-            'SELECT product, available FROM lakey.shopping ) '
-            'AS q WHERE RAND() >= 0.9')
-
-    #
     # GET_SAMPLE
     #
     @override_settings(
